Refactored codebase and improved modularity
